+++ lib/input/s3.go
@@ -71,7 +71,13 @@ This input adds the following metadata fields to each message:
 
 ` + "```" + `
 - s3_key
-- All existing file metadata (only when NOT using download manager)
+- s3_bucket
+- s3_last_modified_unix \*
+- s3_last_modified (RFC3339) \*
+- s3_content_type \*
+- All user defined metadata \*
+
+\* Only added when NOT using download manager
 ` + "```" + `
 
 You can access these metadata fields using